refactor: moved out UserRoleRow
